% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimodal-wrapper.R
\name{multimodal_sPLS_wrapper}
\alias{multimodal_sPLS_wrapper}
\title{MuliModalSparsePLS(DA)}
\usage{
multimodal_sPLS_wrapper(
  mae,
  study_assays,
  lineages = NULL,
  stages = NULL,
  ncomp = 2,
  scale = TRUE,
  design = "null",
  DA = NULL,
  keepX = NULL,
  save = TRUE
)
}
\arguments{
\item{mae}{A MultiAssayExperiment object}

\item{study_assays}{study assays. first must be 'rna' for unsupervised analysis}

\item{lineages}{Character, lineages to keep, from unique(colData(mae)$lineage10x_2)}

\item{stages}{Character, stages to keep, from unique(colData(mae)$stage)}

\item{ncomp}{Integer, number of components}

\item{scale}{Logical, whether to scale the features}

\item{DA}{NULL for unsupervised, or from names(colData(mae)) for supervised (PLSDA)}
}
\description{
MuliModalSparsePLS(DA)
}
