% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbindListWithNames}
\alias{rbindListWithNames}
\title{rbind a list of arrays}
\usage{
rbindListWithNames(lst, new_col = "dataset")
}
\arguments{
\item{lst}{A named list of arrays}

\item{new_col}{Character, the name of the new column which adds the list names
to the resulting array}
}
\description{
rbind a list of arrays
}
